bot/requirements.txt
pyrogram
tgcrypto
asyncio
aiohttp
aiofiles
python-dotenv
